% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{diet_summary}
\alias{diet_summary}
\title{Diet summaries from the SCAR Southern Ocean Diet and Energetics Database}
\usage{
diet_summary(
  x,
  summary_type = "prey",
  minimum_importance = 0,
  treat_trace_values_as = 0
)
}
\arguments{
\item{x}{data.frame: diet data, as returned by \code{so_diet}}

\item{summary_type}{string: either "predators" (report the predators in the data) or "prey" (report the prey items)}

\item{minimum_importance}{numeric: ignore records with dietary importance less than this threshold}

\item{treat_trace_values_as}{numeric: what numeric value to use for a dietary item recorded as "trace"}
}
\value{
data.frame with columns:
\itemize{
   \item{N_fraction_diet_by_weight - number of diet observations where fraction of diet by weight was quantified}
   \item{fraction_diet_by_weight - mean fraction of diet by weight}
   \item{N_fraction_occurrence - number of diet observations where fraction of occurrence quantified}
   \item{fraction_occurrence - mean fraction of occurrence}
   \item{N_fraction_diet_by_prey_items - number of diet observations where fraction of diet by number of prey items was quantified}
   \item{fraction_diet_by_prey_items - mean fraction of diet by number of prey items}
}
}
\description{
Given a diet data.frame, the prey or predators are aggregated to group level and dietary importance values reported.
Note that dietary importance values are currently calculated by unweighted averaging across studies (so that e.g. a
study of 100 individuals will carry the same weight as a study of one individual).
}
\examples{
\dontrun{
  x <- so_diet()

  ## summary of what Electrona carlsbergi eats
  x \%>\% filter_by_predator_name("Electrona carlsbergi") \%>\%
  diet_summary(summary_type = "prey")

  ## summary of what eats Electrona carlsbergi
  x \%>\% filter_by_prey_name("Electrona carlsbergi") \%>\%
  diet_summary(summary_type = "predators")
}

}
