% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_name}
\alias{filter_by_name}
\title{Filter data by predator, prey, or taxon name}
\usage{
filter_by_name(x, name, name_type)
}
\arguments{
\item{x}{data.frame: diet data as returned by e.g. \code{so_diet} or \code{so_isotopes}}

\item{name}{character: vector of one or more names to match on}

\item{name_type}{string: one of "predator", "prey", "predatorprey" (to match on either predator names or prey names), or "taxon". If missing will default to predator names for diet data, and taxon names for isotope data}
}
\value{
data.frame
}
\description{
A convenience function that matches on the species and/or group names.
}
\examples{
\dontrun{
  x <- so_isotopes()
  x \%>\% filter_by_name(c("Electrona","Gymnoscopelus"),"taxon")
}

}
\seealso{
\code{\link{so_diet}}, \code{\link{so_isotopes}}
}

