% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_by_importance}
\alias{replace_by_importance}
\title{Replace diet data entries below a given importance threshold}
\usage{
replace_by_importance(x, threshold, replace_with = NA,
  which_measures = "all")
}
\arguments{
\item{x}{data.frame: diet data, as returned by \code{so_diet}}

\item{threshold}{numeric: replace entries below this threshold}

\item{replace_with}{numeric: value to replace with}

\item{which_measures}{string: one or more of "fraction_diet_by_weight", "fraction_diet_by_prey_items", "fraction_occurrence", or "all" (shorthand for all three)}
}
\value{
data.frame
}
\description{
Replace diet data entries below a given importance threshold
}
\examples{
\dontrun{
  x <- so_diet()
  ## discard entries representing less than 10\% of diet
  x <- x \%>\% replace_by_importance(0.1)
}

}
\seealso{
\code{\link{so_diet}} \code{\link{filter_by_importance}}
}
