% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_trace_values}
\alias{replace_trace_values}
\title{Replace "trace" values in diet data}
\usage{
replace_trace_values(x, replace_with = 0)
}
\arguments{
\item{x}{data.frame: diet data, as returned by \code{so_diet}}

\item{replace_with}{numeric: value to use as a replacement for "trace" values}
}
\value{
data.frame
}
\description{
Some diet studies record small amounts of a prey item as "trace".
These are encoded in the diet database using the value -999.
This function replaces those entries with the nominated numeric value.
}
\examples{
\dontrun{
  x <- so_diet() \%>\% replace_trace_values()
}

}
\seealso{
\code{\link{so_diet}}
}

