% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{diet_summary}
\alias{diet_summary}
\title{Diet summaries from the SCAR Southern Ocean Diet and Energetics Database}
\usage{
diet_summary(x, predator_name, prey_name, minimum_importance = 0,
  treat_trace_values_as = 0)
}
\arguments{
\item{x}{data.frame: diet data, as returned by \code{so_diet}}

\item{predator_name}{string: name of predator taxon}

\item{prey_name}{string: name of prey taxon}

\item{minimum_importance}{numeric: ignore records with dietary importance less than this threshold}

\item{treat_trace_values_as}{numeric: what numeric value to use for a dietary item recorded as "trace"}
}
\value{
data.frame with columns:
\itemize{
   \item{fraction_diet_by_weight - mean fraction of diet by weight}
   \item{N_fraction_diet_by_weight - number of diet observations where fraction of diet by weight was quantified}
   \item{fraction_occurrence - mean fraction of occurrence}
   \item{N_fraction_occurrence - number of diet observations where fraction of occurrence quantified}
   \item{fraction_diet_by_prey_items - mean fraction of diet by number of prey items}
   \item{N_fraction_diet_by_prey_items - number of diet observations where fraction of diet by number of prey items was quantified}
}
}
\description{
Given a predator (prey) name, the prey (predators) are aggregated to group level.
}
\examples{
\dontrun{
  library(dplyr)
  x <- so_diet()

  ## summary of what Electrona carlsbergi eats
  x \%>\% diet_summary(predator_name="Electrona carlsbergi")

  ## summary of what eats Electrona carlsbergi
  x \%>\% diet_summary(prey_name="Electrona carlsbergi")
}

}

