% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{so_diet}
\alias{so_diet}
\alias{so_dna_diet}
\alias{so_isotopes}
\alias{so_energetics}
\alias{so_lipids}
\alias{so_sources}
\title{Load data from the SCAR Southern Ocean Diet and Energetics database}
\usage{
so_diet(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_dna_diet(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_isotopes(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE,
  format = "wide")

so_energetics(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_lipids(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_sources(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)
}
\arguments{
\item{method}{string: "get" (fetch the data via a web GET call) or "direct" (direct database connection, for internal AAD use only)}

\item{cache_directory}{string: (optional) cache the data locally in this directory, so that they can be used offline later. Values can be "session" (a per-session temporary directory will be used, default), "persistent" (the directory returned by \code{rappdirs::user_cache_dir} will be used), or a string giving the path to the directory to use. Use \code{NULL} for no caching. An attempt will be made to create the cache directory if it does not exist. A warning will be given if a cached copy of the data file exists and is more than 30 days old. Use \code{refresh_cache = TRUE} to refresh the cached data if necessary}

\item{refresh_cache}{logical: if TRUE, and data already exist in the cache_directory, they will be refreshed. If FALSE, the cached data will be used}

\item{public_only}{logical: only applicable to \code{method} "direct"}

\item{verbose}{logical: show progress messages?}

\item{format}{string: (for \code{so_isotopes} only) if "wide", the data will be formatted with one row per record and multiple measurements (of different isotopes) per row. If \code{format} is "mv", the data will be in measurement-value (long) format with multiple rows per original record, split so that each different isotope measurement appears in its own row. See Details for future changes to the default value of this parameter}
}
\value{
data.frame
}
\description{
Load data from the SCAR Southern Ocean Diet and Energetics database. Data will be fetched from the remote server and optionally cached locally, or fetched from the local cache, depending on the arguments passed. Note that \code{so_isotopes} now has a \code{format} parameter - see Details below.
}
\details{
The \code{format} parameter was introduced to the \code{so_isotopes} function in package version 0.4.0. The current default \code{format} is "wide", in which case the data will be formatted with one row per record and multiple measurements (of different isotopes) per row. If \code{format} is "mv", the data will be in measurement-value (long) format with multiple rows per original record, split so that each different isotope measurement appears in its own row. Note that "mv" will become the default (and possibly only) option in a later release. Currently, \code{record_id} values in measurement-value format are not unique (they follow the \code{record_id} values from the wide format). The \code{record_id} values in measurement-value format are likely to change in a future release.
}
\examples{
\dontrun{
  ## conventional diet data
  x <- so_diet()
  subset(x, predator_name == "Electrona antarctica")

  ## DNA diet data
  x <- so_dna_diet()
  subset(x, predator_name == "Thalassarche melanophris")

  ## stable isotopes in wide format
  x <- so_isotopes()

  ## stable isotopes in measurement-value (long) format
  x <- so_isotopes(format = "mv")

  ## energetics data
  x <- so_energetics()

  ## lipids data
  x <- so_lipids()
}
}
\references{
\url{http://data.aad.gov.au/trophic/}
}
