% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sohungry.R
\docType{package}
\name{sohungry}
\alias{sohungry}
\title{\pkg{sohungry}}
\description{
Provides access to data from the SCAR Southern Ocean Diet and Energetics Database.
}
\references{
\url{http://data.aad.gov.au/trophic}
}
